Adding in portage-2.0.47-r12 for testing, and cleaning out the cruft.

+++ sys-apps/portage/portage-2.0.47-r12.ebuild
@@ -1,5 +1,5 @@
 # Distributed under the terms of the GNU General Public License v2 
-# $Header: /var/cvsroot/gentoo-x86/sys-apps/portage/Attic/portage-2.0.46-r12.ebuild,v 1.5 2003/03/11 20:50:08 seemant Exp $
+# $Header: /var/cvsroot/gentoo-x86/sys-apps/portage/Attic/portage-2.0.47-r12.ebuild,v 1.1 2003/03/28 05:23:40 carpaski Exp $
 
 IUSE="build"
 
@@ -10,9 +10,10 @@ export SANDBOX_DISABLED="1"
 S=${WORKDIR}/${PF}
 SLOT="0"
 DESCRIPTION="Portage ports system"
-SRC_URI="mirror://gentoo/${PF}.tar.bz2 http://gentoo.twobit.net/portage/${PF}.tar.bz2"
+SRC_URI="http://gentoo.twobit.net/portage/${PF}.tar.bz2 mirror://gentoo/${PF}.tar.bz2"
 HOMEPAGE="http://www.gentoo.org"
-KEYWORDS="alpha hppa mips ppc sparc x86"
+KEYWORDS="alpha arm hppa mips ppc sparc x86"
+#KEYWORDS="~alpha ~arm ~hppa ~mips ~ppc ~sparc ~x86"
 LICENSE="GPL-2"
 RDEPEND="!build? ( >=sys-apps/fileutils-4.1.8 dev-python/python-fchksum >=dev-lang/python-2.2.1 sys-apps/debianutils >=app-shells/bash-2.05a )"
 
@@ -42,10 +43,6 @@ src_compile() {
 	else
 		make || die
 	fi
-	
-	cp ${S}/bin/emerge ${S}/bin/emerge.orig
-	sed 's:--timeout=60:--timeout=180:' < ${S}/bin/emerge.orig > ${S}/bin/emerge.new && mv ${S}/bin/emerge.new ${S}/bin/emerge
-	rm -f ${S}/bin/emerge.orig ${S}/bin/emerge.new
 }
 
 src_install() {
@@ -56,6 +53,10 @@ src_install() {
 		alpha )
 		newins make.globals.alpha make.globals
 		newins make.conf.alpha make.conf
+		;;
+		arm )
+		newins make.globals.arm make.globals
+		newins make.conf.arm make.conf
 		;;
 		hppa )
 		newins make.globals.hppa make.globals
@@ -77,7 +78,9 @@ src_install() {
 		doins make.globals make.conf
 		;;
 	esac
-	doins etc-update.conf
+	use build && rm -f ${D}/etc/make.conf
+	
+	doins etc-update.conf dispatch-conf.conf
 	#python modules
 	cd ${S}/src/python-missingos
 	chmod +x setup.py
@@ -116,6 +119,8 @@ src_install() {
 	dosym ../lib/portage/bin/g-cpan.pl /usr/bin/g-cpan.pl
 	dosym ../lib/portage/bin/quickpkg /usr/bin/quickpkg
 	dosym ../lib/portage/bin/regenworld /usr/sbin/regenworld
+	dosym ../lib/portage/bin/emerge-webrsync /usr/sbin/emerge-webrsync
+	dosym ../lib/portage/bin/dispatch-conf /usr/sbin/dispatch-conf
 
 	dosym newins /usr/lib/portage/bin/donewins
 	
@@ -132,25 +137,6 @@ src_install() {
 }
 
 pkg_postinst() {
-	if [ ! -e ${ROOT}/etc/make.profile ]
-	then
-		cd ${ROOT}/etc
-		case "$ARCH" in
-		ppc )
-			ln -sf ../usr/portage/profiles/default-ppc-1.0 make.profile
-			;;
-		sparc )
-			ln -sf ../usr/portage/profiles/default-sparc-1.0 make.profile
-			;;
-		sparc64 )
-			ln -sf ../usr/portage/profiles/default-sparc64-1.0 make.profile
-			;;
-		x86 )
-			ln -sf ../usr/portage/profiles/default-1.0 make.profile
-			;;
-		esac
-
-	fi
 	local x
 
 	#disable global sandbox if it's active (it's been deprecated)
@@ -174,18 +160,54 @@ pkg_postinst() {
 		rm -f /var/cache/edb/mtimes
 	fi
 
-	# Kill the existing counter and generate a new one.
-	echo -n "Recalculating the counter... "
-	mv /var/cache/edb/counter /var/cache/edb/counter.old
-	python -c 'import portage; portage.counter_tick_core("/")' &>/dev/null
-	if [ -f /var/cache/edb/counter ] ; then
-		echo "Counter updated successfully."
-		rm -f /var/cache/edb/counter.old
-	else
-		echo "FAILED to update counter."
-		echo "!!! This is a problem."
-		mv /var/cache/edb/counter.old /var/cache/edb/counter
-	fi
+	echo
+	einfo "NOTICE: PLEASE update your make.globals. All user changes to variables"
+	einfo "in make.globals should be placed in make.conf. DO NOT MODIFY make.globals."
+	echo
+	einfo "NOTICE: The wheel group requirement for non-root users has been changed to"
+	einfo "group portage. Group portage must be a valid group for user to use portage."
+	echo
+	einfo "Feature additions are noted in help and make.conf descriptions. Update"
+	einfo "them using 'etc-update' please. Maintaining current configs for portage"
+	einfo "and other system packages is fairly important for the continued health"
+	einfo "of your system."
+	echo
+	einfo "A worldfile rebuilding script is available to regenerate entries that"
+	einfo "should be in your worldfile but were removed by a recently discovered"
+	einfo "'-e bug' or if you deleted it: run 'regenworld' as root."
+	echo
+	einfo "The 2.0.47 line of portages contains an optional userpriv mode that"
+	einfo "enables portage to drop root privleges and run as a normal user. It is"
+	einfo "enabled via FEATURES by adding userpriv."
+	echo
+	einfo "Please 'emerge sync' after merging portage to update some permissions."
+	echo
+	if [ -z $PORTAGE_TEST ]; then
+		echo -ne "\a" ; sleep 1 ; echo -ne "\a" ; sleep 1 ; echo -ne "\a" ; sleep 1
+		echo -ne "\a" ; sleep 1 ; echo -ne "\a" ; sleep 1 ; echo -ne "\a" ; sleep 1
+		sleep 5
+
+		# Kill the existing counter and generate a new one.
+		echo -n "Recalculating the counter... "
+		mv /var/cache/edb/counter /var/cache/edb/counter.old
+		python -c 'import portage; portage.counter_tick_core("/")' &>/dev/null
+		if [ -f /var/cache/edb/counter ] ; then
+			echo "Counter updated successfully."
+			rm -f /var/cache/edb/counter.old
+		else
+			echo "FAILED to update counter."
+			echo "!!! This is a problem."
+			mv /var/cache/edb/counter.old /var/cache/edb/counter
+		fi
+
+		# Changes in the size of auxdbkeys can cause aux_get() problems.
+		echo -n ">>> Clearing invalid entries in dependancy cache..."
+		cd ${ROOT}var/cache/edb/dep
+		# 2>&1 >/dev/null <---- Kills stdout, replaces it with stderr
+		AUXDBKEYLEN="$(python -c 'import portage,sys; sys.stderr.write(str(len(portage.auxdbkeys)))' 2>&1 >/dev/null)"
+		find ${ROOT}var/cache/edb/dep -type f -exec wc -l {} \; | egrep -v "^ *${AUXDBKEYLEN}" | sed 's:^ \+[0-9]\+ \+\([^ ]\+\)$:\1:' 2>/dev/null | xargs -n 50 -r rm -f
+		echo " ...done!"
+	fi # PORTAGE_TESTING
 
 	#fix cache (could contain staleness)
 	if [ ! -d ${ROOT}var/cache/edb/dep ]
@@ -209,8 +231,9 @@ pkg_postinst() {
 			[ -d ${ROOT}var/db/pkg/virtual ] && rm -rf ${ROOT}var/db/pkg/virtual
 		fi
 		install -d -m0755 ${ROOT}var/cache/edb
-		install -d -m2775 -o root -g wheel ${ROOT}var/cache/edb/dep
+		install -d -m2775 -o root -g portage ${ROOT}var/cache/edb/dep
 	fi
+
 	rm -f ${ROOT}usr/lib/python2.2/site-packages/portage.py[co]
 	rm -f ${ROOT}usr/lib/python2.2/site-packages/output.py[co]
 	chmod 2775 ${ROOT}var/cache/edb/dep ${ROOT}var/cache/edb/dep/*
@@ -224,43 +247,8 @@ pkg_postinst() {
 	python -c "import py_compile; py_compile.compile('${ROOT}usr/lib/portage/bin/emergehelp.py')" || die
 	python -O -c "import py_compile; py_compile.compile('${ROOT}usr/lib/portage/bin/emergehelp.py')" || die
 
-	# Changes in the size of auxdbkeys can cause aux_get() problems.
-	echo -n ">>> Clearing invalid entries in dependancy cache..."
-	cd ${ROOT}var/cache/edb/dep
-	# 2>&1 >/dev/null <---- Kills stdout, replaces it with stderr
-	AUXDBKEYLEN="$(python -c 'import portage,sys; sys.stderr.write(str(len(portage.auxdbkeys)))' 2>&1 >/dev/null)"
-	find ${ROOT}var/cache/edb/dep -type f -exec wc -l {} \; | egrep -v "^ *${AUXDBKEYLEN}" | sed 's:^ \+[0-9]\+ \+\([^ ]\+\)$:\1:' 2>/dev/null | xargs -n 50 -r rm -f
-	echo " ...done!"
-
-	echo
-	einfo "NOTICE: PLEASE update your make.globals. All user changes to variables"
-	einfo "in make.globals should be placed in make.conf. DO NOT MODIFY make.globals."
-	einfo "AUTOCLEAN's default has been changed to 'yes' to ensure that libraries are"
-	einfo "treated properly during merges. NOT updating make.globals may result in you"
-	einfo "experiencing missing symlinks, failed compiles, and the inability to log in"
-	einfo "to your system. Running 'ldconfig' should fix the majority of these problems,"
-	einfo "but you may need to boot from a gentoo cd and execute the following:"
-	einfo "chroot /mnt/gentoo /sbin/ldconfig"
-	echo
-	einfo "Feature additions are noted in help and make.conf descriptions. Update"
-	einfo "them using 'etc-update' please. Maintaining current configs for portage"
-	einfo "and other system packages is fairly important for the continued health"
-	einfo "of your system."
-	echo
-	einfo "A worldfile rebuilding script is available to regenerate entries that"
-	einfo "should be in your worldfile but were removed by a recently discovered"
-	einfo "'-e bug' or if you deleted it: run 'regenworld' as root."
-	echo
-
-	echo -ne "\a" ; sleep 1 ; echo -ne "\a" ; sleep 1 ; echo -ne "\a" ; sleep 1
-	echo -ne "\a" ; sleep 1 ; echo -ne "\a" ; sleep 1 ; echo -ne "\a" ; sleep 1
-	echo -ne "\a" ; sleep 1 ; echo -ne "\a" ; sleep 1 ; echo -ne "\a" ; sleep 1
-	sleep 6
-
-	#OLDPV=$(< ${WORKDIR}/previous-version)
-	#if compare_pver "$OLDPV" "$(get_portver)"; then
-	#	echo -ne "\a"; sleep 1; echo -ne "\a"; sleep 1; echo -ne "\a"; sleep 1
-	#	echo -ne "\a"; sleep 1; echo -ne "\a"; sleep 1; echo -ne "\a"; sleep 1
-	#	echo -ne "\a"; sleep 1; echo -ne "\a"; sleep 1; echo -ne "\a"; sleep 1
-	#fi	
+	if has ccache $FEATURES && has userpriv $FEATURES; then
+		chown -R portage:portage /var/tmp/ccache &> /dev/null
+		chmod -R g+rws /var/tmp/ccache &>/dev/null
+	fi
 }
